requirements.txt
flask
flask_sqlalchemy
pyodbc
azure-storage